requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.1
compound-word-splitter==0.4
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
datasets==2.16.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
emoji==2.11.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
GPUtil==1.4.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
inflect==7.0.0
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langcodes==3.3.0
langid==1.1.6
language-tool-python==2.8
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nest-asyncio==1.5.8
networkx==3.3
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
platformdirs==4.1.0
preshed==3.0.9
prompt-toolkit==3.0.42
protobuf==5.26.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pydantic==2.5.3
pydantic_core==2.14.6
pyenchant==3.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
setuptools==69.0.3
six==1.16.0
smart-open==6.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
stanza==1.8.1
sympy==1.12
tabulate==0.9.0
thinc==8.2.2
tokenizers==0.15.0
toml==0.10.2
torch==2.2.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.0.dev0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
wheel==0.42.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
